requirements.txt
google-cloud-pubsub>=2.6.1